pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ooouno-ex"
version = "2.0.0"
description = "Examples for LibreOffice python"
license = "MIT"
readme = "README.md"
authors = [
    ":Barry-Thomas-Paul: Moss <bigbytetech@gmail.com>"
]
homepage = "https://github.com/Amourspirit/python-ooouno-ex"
documentation = "https://github.com/Amourspirit/python-ooouno-ex"
repository = "https://github.com/Amourspirit/python-ooouno-ex"


[tool.poetry.dependencies]
python = "^3.8"
oooscript = ">=1.1.3"
types-access2base = ">=1.0.2"
types-scriptforge = ">=2.0.1"
types-unopy = ">=1.0.3"
unotools = ">=0.3.3"
verr = ">=1.1.2"
kwargshelper = ">=2.7.1"
ooo-dev-tools-gui-win = { version = ">=0.3.1", markers = "sys_platform == 'win32'" }
ooo-dev-tools = ">=0.26.0"

[tool.poetry.group.dev.dependencies]
black = ">=22.12.0"
oooenv = ">=0.2.2"

[tool.ruff]
ignore=["E501"]